requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
blinker=1.8.2=pyhd8ed1ab_0
bzip2=1.0.8=h93a5062_5
ca-certificates=2024.2.2=hf0a4a13_0
cachetools=4.2.4=pypi_0
certifi=2024.2.2=pypi_0
chardet=3.0.4=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=unix_pyh707e725_0
filelock=3.14.0=pypi_0
flask=3.0.3=pyhd8ed1ab_0
fsspec=2024.3.1=pypi_0
google-api-core=1.34.1=pypi_0
google-auth=1.35.0=pypi_0
google-cloud-core=1.7.3=pypi_0
google-cloud-translate=2.0.1=pypi_0
googleapis-common-protos=1.63.0=pypi_0
googletrans=3.1.0a0=pypi_0
grpcio=1.63.0=pypi_0
grpcio-status=1.48.2=pypi_0
h11=0.9.0=pypi_0
h2=3.2.0=pypi_0
hpack=3.0.0=pypi_0
hstspreload=2024.5.1=pypi_0
httpcore=0.9.1=pypi_0
httpx=0.13.3=pypi_0
hyperframe=5.2.0=pypi_0
idna=2.10=pypi_0
importlib-metadata=7.1.0=pyha770c72_0
itsdangerous=2.2.0=pyhd8ed1ab_0
jinja2=3.1.4=pyhd8ed1ab_0
libexpat=2.6.2=hebf3989_0
libffi=3.4.2=h3422bc3_5
libsqlite=3.45.3=h091b4b1_0
libzlib=1.2.13=h53f4e23_5
llvmlite=0.42.0=pypi_0
markupsafe=2.1.5=py312he37b823_0
more-itertools=10.2.0=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.5=hb89a1cb_0
networkx=3.3=pypi_0
numba=0.59.1=pypi_0
numpy=1.26.4=pypi_0
openai-whisper=20231117=pypi_0
openssl=3.3.0=h0d3ecfb_0
pip=24.0=pyhd8ed1ab_0
protobuf=3.20.3=pypi_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
python=3.12.3=h4a7b5fc_0_cpython
python_abi=3.12=4_cp312
readline=8.2=h92ec313_1
regex=2024.5.10=pypi_0
requests=2.31.0=pypi_0
rfc3986=1.5.0=pypi_0
rsa=4.9=pypi_0
setuptools=69.5.1=pyhd8ed1ab_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
sympy=1.12=pypi_0
tiktoken=0.6.0=pypi_0
tk=8.6.13=h5083fa2_1
torch=2.3.0=pypi_0
tqdm=4.66.4=pypi_0
typing-extensions=4.11.0=pypi_0
tzdata=2024a=h0c530f3_0
urllib3=2.2.1=pypi_0
werkzeug=3.0.3=pyhd8ed1ab_0
wheel=0.43.0=pyhd8ed1ab_1
xz=5.2.6=h57fd34a_0
zipp=3.17.0=pyhd8ed1ab_0
